requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
Deprecated==1.2.18
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
freezegun==1.5.5
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
install==1.3.5
isort==7.0.0
Jinja2==3.1.6
jiter==0.11.1
limits==5.6.0
logging==0.4.9.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openai==2.5.0
packaging==25.0
pluggy==1.6.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
sentry-sdk==2.42.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
